requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
numpy==2.0.0
pyepics==3.5.6
pyparsing==3.1.2
requests==2.32.3
urllib3==2.2.2
